Update styles.scss
"Developer Login" in the header can wrap in an ugly way if the screen is between 981-1130px wide. This breaks the look/feel of the button, and shifts the rest of the page content downwards. This CSS change does not allow the UI to break between "Developer" and "Login". I can't see anywhere else where applying this to the `.button` class has any effect.