Remove beta URL segment for Castamatic
I am new to the Postcast Index and Castamatic.

I wanted to try out the Value4Value concept and Castamatic happened to be the first app I downloaded. The link pointed to the TestFlight beta which confused me at first. I downloaded the main Castamatic app, not the beta, and it seem to support all the features from the `supportedElements` so I suppose the beta link is no longer needed?