Apps page filter expansion based on state
I did this change while investigating if I could implement
https://github.com/Podcastindex-org/web-ui/issues/139.

The arrow on the Filters was being changed directly by
manipulating the element text.

I replaced it by a state that indicates wheter it is
expanded or not.

I also wanted to animate solely based on state, but
quickly discovered how tricky that was, thus I
added a comment to an article explaining why the
`collapseSection` and `expandSection` mumbo jumbo is
actually necessary.