Refactoring: Filtered apps based on state
The filter buttons on the Apps UI were not drawn or updated based
on state. This commit changes that.

Since the logic for all filter types is the same, to avoid
multiple state and state setter "ifs", I hidden that
behind functions that get the filter state and
filter state setters.

I also separated the Clear button from the list of tags.