Apps page: delegate setting filtered apps to FilterTags
This commit completely delegates the setting of filtered apps
to the FilterTags component, as a preparation to have
initial filtering based on query parameters.

It is important to remove the setting of filtered apps from
the useEffect that reacts to changes on the apps, as it would
overwrite changes comming from the FilterTags as a reaction
to changes on the apps.