Apps Page: Update URL Search Parameters on filter change This commit introduces the featue of updatading the URL Search Parameters when the filters are changed.
This also means that now when a user clicks a filter
button, the state is not directly updated, but
only indirectly as a reaction to the change in the
URL Search Params.

Updating both the URL Search Parameters and the state
was tested, but not only it was unnecessary, it caused
undesired side effects, causing sometimes infinite
render loops.

It also fixes a an issue introduced with:
https://github.com/Podcastindex-org/web-ui/pull/142

The Clear button was not working anymore.