Apps: Added filter for open source apps
This commit adds the possibility for filtering for Open Source
apps in the App Type filter.

I added the `open source` type to sever apps that I could
verify are open source.

It would be interesting to have an additional field to
link directly to the source code, and show this in
the UI somehow.