Issue #213 - UI - Initial support for comments
This commit introduces initial support for showing comments.
It relies on the API also including socialInteracts and on a
new API on the server being implemented to fetch the comments.

This is not yet ready, I have implemented some mock code on
the server side that is not included in this commit.

The front-end code expects the comments API to respond with
the same format as podverse's
/api/v1/episode/{episodeid}/proxy/activity-pub

Related issues:
- resolves #213
- https://github.com/Podcastindex-org/web-ui/issues/213